streaming/data_ingestion/pyproject.toml
[tool.poetry]
name = "data_ingestion"
version = "0.1.0"
description = ""
authors = ["William Chong <william@inorbit.com>"]
readme = "README.md"
packages = [{include = "data_ingestion"}]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
